ci: skip deleted packages during gc

+++ cleanup-images.sh
@@ -15,6 +15,9 @@ github_paged() {
 }
 
 github_paged 'https://api.github.com/user/packages?package_type=container' '.[].url' | while read -r package_url; do
+    if [[ "${package_url}" == */deleted_* ]]; then
+        continue
+    fi
     github_paged "${package_url}/versions?package_type=container" '.[]|select(.metadata.container.tags | length == 0)|.url' | while read -r version_url; do
         github "${version_url}" -X DELETE 
     done
